Changed size_t to unsigned int for 64 bit systems

+++ src/libzzub/ccm.cpp
@@ -1075,7 +1075,7 @@ namespace zzub {
 
   struct DecodedFrame {
     void* buffer;
-    size_t bytes;
+    unsigned int bytes;
   };
 
   struct DecoderInfo {
